requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
channels==4.3.1
channels-redis==4.2.0
redis==6.4.0
python-decouple==3.8
django-health-check==3.17.0
django-redis==5.4.0
gunicorn==21.2.0
uvicorn==0.27.0
aiohttp==3.11.10   